Write the main client connection method body

+++ lttnetworking/src/messages/hello.rs
@@ -14,5 +14,5 @@ pub struct ServerHello {
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ServerInfo {
-    pub max_sub_connections: usize,
+    pub max_sub_connections: u64,
 }
